Restructure installer a bit
