+++ discord/installer/lib/addon.js
