+++ discord/installer/lib/integration.js
