+++ discord/installer/lib/paths.js
